Apple: extend the refresh shield to detail-view flag writes
The pending-write shield only covered flag toggles issued by the list
view model. A \Seen/\Flagged toggle from the detail view took a different
path - the detail VM does the STORE and the list only mirrors the flag
via AppState - so a refresh landing mid-write could still revert it.

Bracket the detail VM's flag writes through a shared, folder-keyed
registry: AppState.pendingFlagWriteUIDs, populated via setFlagWrite(...)
before the STORE and cleared (defer) when it resolves. The detail VM
exposes the bracket as an onFlagWriteInFlight callback (same decoupling
as onFlagChanged); MessageDetailView wires it to AppState. The list's
shieldFetched now consults this registry (scoped to its folder.path)
alongside its own pendingFlagUIDs. The registry is read directly at merge
time rather than mirrored via onChange, so there's no view-update hop
between the optimistic flip and the shield taking effect. Folder-keying
prevents a UID collision across mailboxes from mis-shielding a row.

Extract the \Seen/\Flagged toggles and the mark-as-read scheduler into
MessageDetailViewModel+Flags.swift to keep the type body under the
SwiftLint cap; client/preferences/pendingMarkAsReadTask relax to internal
for the sibling extension, matching MessageListViewModel's split.
